Decompose name filtering into hook
