+++ src/pages/OrderStock.js
@@ -29,7 +29,10 @@ import { LANDING, ORDER_FORM } from '../constants/Routes';
 import { addPersonInfo } from '../api/People';
 import { updateStockCount } from '../api/Stock';
 import { auth } from '../firebase';
+
+/** Custom hooks */
 import useStock from '../hooks/useStock';
+import useNameSearch from '../hooks/useNameSearch';
 
 /**
  * Allows user to order stock items only after they've filled out form
@@ -42,7 +45,7 @@ function OrderStock() {
 
   const [stock, loading, error, getStock] = useStock();
   const [language, setLanguage] = useState(LANGUAGES[0]);
-  const [nameQuery, setNameQuery] = useState('');
+  const [setNameQuery, getFilteredStockItems] = useNameSearch(stock, language.tag);
   const [personInfo] = useState(
     fromForm
       ? location.state.personInfo
@@ -178,28 +181,6 @@ function OrderStock() {
     }
   }
 
-  /**
-   * Returns filtered stock array based on search queries
-   * Name query: allows if EITHER English or translated name includes query
-   */
-  function getFilteredStockItems() {
-    return stock.filter((item) => {
-      // Check if query is included in English name
-      const inEnglishName = item.name.toLowerCase().includes(nameQuery);
-
-      let inTranslatedName = false;
-
-      // Check first if translations have loaded
-      let translations = item.translations;
-      if (Boolean(translations) && translations[language.tag] !== undefined) {
-        // Check if query is included in translated name
-        inTranslatedName = item.translations[language.tag]
-          .toLowerCase()
-          .includes(nameQuery);
-      }
-      return inEnglishName || inTranslatedName;
-    });
-  }
 
   const classes = useStyles();
   return (
